+++ ChessGame/model/Knight.cs
@@ -4,9 +4,9 @@ namespace ChessGame.model
 {
     public class Knight : ChessPiece
     {
-        private readonly int boardSize;
+        private readonly byte boardSize;
 
-        public Knight(ColorType color, int size) : base(color, size)
+        public Knight(ColorType color, byte size) : base(color, size)
         {
             boardSize = BoardSize;
             Movement = MovementType.singleSquare;
@@ -31,7 +31,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetLowerLeftDiagonal()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX - 1, CurrentY + 2),
+              new Position((byte)(CurrentX - 1), (byte)(CurrentY + 2)),
               CurrentX > 0 && CurrentY < boardSize - 1
             );
         }
@@ -39,7 +39,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetLeftLine()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX - 2, CurrentY + 1),
+              new Position((byte)(CurrentX - 2), (byte)(CurrentY + 1)),
               CurrentX > 1 && CurrentY < boardSize
             );
         }
@@ -47,7 +47,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetUpperLeftDiagonal()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX - 1, CurrentY - 2),
+              new Position((byte)(CurrentX - 1), (byte)(CurrentY - 2)),
               CurrentX > 0 && CurrentY > 1
             );
         }
@@ -55,7 +55,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetUpperColumn()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX - 2, CurrentY - 1),
+              new Position((byte)(CurrentX - 2), (byte)(CurrentY - 1)),
               CurrentX > 1 && CurrentY > 0
             );
         }
@@ -63,7 +63,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetLowerColumn()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX + 1, CurrentY + 2),
+              new Position((byte)(CurrentX + 1), (byte)(CurrentY + 2)),
               CurrentX < boardSize && CurrentY < boardSize - 1
             );
         }
@@ -71,7 +71,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetLowerRightDiagonal()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX + 2, CurrentY + 1),
+              new Position((byte)(CurrentX + 2), (byte)(CurrentY + 1)),
               CurrentX < boardSize - 1 && CurrentY < boardSize
             );
         }
@@ -79,7 +79,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetUpperRightDiagonal()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX + 1, CurrentY - 2),
+              new Position((byte)(CurrentX + 1), (byte)(CurrentY - 2)),
               CurrentX < boardSize && CurrentY > 1
             );
         }
@@ -87,7 +87,7 @@ namespace ChessGame.model
         public override IEnumerable<Position> GetRightLine()
         {
             return AddSingleSquarePositions(
-              new Position(CurrentX + 2, CurrentY - 1),
+              new Position((byte)(CurrentX + 2), (byte)(CurrentY - 1)),
               CurrentX < boardSize - 1 && CurrentY > 0
             );
         }
