+++ ChessGame/view/CustomChessSquare.cs
@@ -27,9 +27,9 @@ namespace ChessGame
             rawPicture = square;
         }
 
-        public int X { get; set; }
+        public byte X { get; set; }
 
-        public int Y { get; set; }
+        public byte Y { get; set; }
 
         public PictureBox GetControl() => rawPicture;
